Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 int main() {
-    int a;
+    float a;
     float c;
     printf("");
-    scanf("%d",&a);
+    scanf("%f",&a);
     c = 3.14*a*a;
     printf("Area: %.2f\n",c);
     return 0;
